Phase 1 milestone: django.setup() succeeds cleanly under Python 3.12
Full app registry (all installed apps, models, URL configs, ready()
hooks) now imports with zero errors and zero warnings under Python
3.12, on the python3-port branch. docker/py3/ is the reusable
validation environment for the rest of this port.

Three categories of fix, roughly in the order they were found:

Third-party packages needing a real version bump (not just a patch -
the previous pins are hard-incompatible with modern Python tooling,
not just old):
- django-assets 0.12 -> 2.0 (imported the removed stdlib `imp` module)
- rq 0.10.0 -> 1.0, django-rq 1.0.1 -> 2.1.0 (rq had a function
  parameter literally named `async=`, a reserved keyword since Python
  3.7 - a hard SyntaxError just importing the module. Picked
  django-rq==2.1.0 deliberately: it's the last release still
  declaring `django>=1.8.0` compatibility, so this doesn't drag
  Django along - an earlier looser-range attempt did exactly that
  before being caught and re-pinned exactly)

Django 1.11 itself (officially only ever supported Python 3.4-3.7)
needing compatibility shims in pootle/syspath_override.py, the
project's existing designated monkey-patch entry point - not
patching Django's source, and removable one hop at a time as Phase
2's ladder proceeds:
- collections.Iterator/Mapping/... aliases, removed in Python 3.10
- gettext.NullTranslations.set_output_charset, removed upstream
- gettext.translation()'s codeset kwarg, removed upstream
- django/utils/six.py (Django's vendored copy) doesn't resolve
  django.utils.six.moves.* under Python 3.12's import system - fixed
  at image-build time instead (docker/py3/patch-django-six.sh) by
  overwriting it with the real `six` package's source, since six's
  own internals key off __name__ at import time and can't be aliased
  after the fact

Our own code (~50 files): the mechanical Python-2-only syntax
(unicode/basestring/xrange/.iteritems() etc., including several
missed by the original grep-based audit - bare `unicode` as a value
not just unicode(...) calls), two implicit relative imports (valid in
Python 2, a hard error since PEP 328), renamed stdlib modules
(urllib2/urlparse -> urllib.parse), removed builtins (unichr, cmp,
reduce moved to functools), a removed contextlib.nested (replaced
with ExitStack for the two sites that needed a variable-length list
of context managers), an invalid `ur"..."` string prefix, mixed
tabs/spaces in 95-outro.conf, and a batch of invalid-escape-sequence
SyntaxWarnings (regex patterns as non-raw strings) that would become
hard failures once the test suite's `filterwarnings = error` policy
applies - fixed as raw strings, each verified by eval round-trip
against the original computed string value before writing, so none
of them silently changed the actual regex being compiled.

Also: setup.py no longer imports pkg_resources (removed from
setuptools by default) or relies on stdlib distutils (removed
entirely in 3.12, replaced by importing setuptools first so its
distutils shim takes effect for the Command base classes still used
here); the vestigial check_pep440_versions() version-floor guard was
removed rather than ported, since any supported toolchain already
clears it trivially.

Not yet run: the actual test suite. django.setup() succeeding proves
the app imports and configures cleanly, not that it behaves
correctly - streams B/C's ~2500 tests haven't been exercised under
Python 3 yet.

+++ pootle/core/schema/mysql.py
@@ -10,7 +10,7 @@ from django.db import connection
 
 
 def type_cast(value):
-    if isinstance(value, long):
+    if isinstance(value, int):
         return int(value)
     return value
